Refactored pushbullet
